提交后台用户管理、后台角色管理、后台权限管理； 提交短信发送服务； 配置文件修改； 增加pom引用

+++ src/main/java/com/farmershao/stock/persistence/mapper/system/SysRoleMapper.java
@@ -15,7 +15,7 @@ import java.util.List;
 public interface SysRoleMapper {
     int insertSelective(SysRole record);
 
-    SysRole selectByPrimaryKey(Integer id);
+    SysRole selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(SysRole record);
