+++ src/main/java/com/farmershao/stock/service/system/SysRoleService.java
@@ -37,7 +37,7 @@ public class SysRoleService {
      * @param id    角色ID
      * @return
      */
-    public SysRole findById(Integer id) {
+    public SysRole findById(Long id) {
         return sysRoleMapper.selectByPrimaryKey(id);
     }
 
@@ -53,7 +53,7 @@ public class SysRoleService {
         String cacheKey = CacheKeyEnum.BACK_PERMISSION_ARRAY.getKey() + roleId;
         String permissions = redisUtil.getValue(cacheKey, String.class);
         if (StringUtils.isEmpty(permissions)) {
-            SysRole sysRole = findById(Integer.valueOf(roleId.trim()));
+            SysRole sysRole = findById(Long.valueOf(roleId.trim()));
             permissions = sysRole.getPermissions();
             redisUtil.setValue(cacheKey, permissions, CacheKeyEnum.BACK_PERMISSION_ARRAY.getExpire());
         }
